+++ src/components/Buttons/Eduplaces/index.tsx
@@ -1,0 +1,3 @@
+import Eduplaces from './Eduplaces';
+
+export { Eduplaces as default };