+++ src/components/UserInfo/UserInfo.tsx
@@ -24,8 +24,8 @@ const UserInfo = () => {
         <div className="flex flex-row items-center min-h-fit gap-2">
           <UserAvatar id={user.hash_id} size={52} />
           <div className="flex-1">
-            <h3 className="mt-1 text-2xl font-bold truncate">{user.username}</h3>
-            <p className="text-sm truncate">{user.displayname}</p>
+            <h3 className="mt-1 text-2xl font-bold truncate">{user.displayname}</h3>
+            <p className="text-sm font-light truncate">{user.realname}</p>
           </div>
         </div>
       )}