+++ src/views/Settings/Config/IdpSyncEntry/index.tsx
@@ -1,0 +1,3 @@
+import IdpSyncEntry from './IdpSyncEntry';
+
+export { IdpSyncEntry as default };