+++ src/views/Settings/IdpSync/index.tsx
@@ -1,0 +1,3 @@
+import IdpSyncView from './IdpSyncView';
+
+export default IdpSyncView;